+++ .bandit
@@ -1,0 +1,4 @@
+[bandit]
+exclude: /tests
+
+tests: B108,B306,B307,B313,B314,B315,B316,B317,B318,B319,B320,B325,B602,B604
