+++ .github/workflows/tests.yaml
@@ -1,4 +1,4 @@
-name: Python package
+name: Tests
 
 on: [push]
