+++ tests/bandit.yaml
@@ -1,0 +1,21 @@
+# https://bandit.readthedocs.io/en/latest/config.html
+
+tests:
+  - B103
+  - B108
+  - B306
+  - B307
+  - B313
+  - B314
+  - B315
+  - B316
+  - B317
+  - B318
+  - B319
+  - B320
+  - B325
+  - B601
+  - B602
+  - B604
+  - B608
+  - B609
