+++ tests/conftest.py
@@ -12,74 +12,43 @@ from PyViCare.PyViCareService import (
     buildSetPropertyUrl,
     readFeature,
 )
-import pytest
-import pytest_asyncio
 from homeassistant.components.vicare.const import DOMAIN
 from homeassistant.core import HomeAssistant
+import pytest
 
 from . import ENTRY_CONFIG
 
-from pytest_homeassistant_custom_component.common import MockConfigEntry
+from tests.common import MockConfigEntry
 
-pytest_plugins = "pytest_homeassistant_custom_component"
-
-# This fixture enables loading custom integrations in all tests.
-# Remove to enable selective use of this fixture
-@pytest.fixture(autouse=True)
-def auto_enable_custom_integrations(enable_custom_integrations):
-    """Automatically enable loading custom integrations in all tests."""
-    yield
-
-
-# This fixture is used to prevent HomeAssistant from attempting to create and dismiss persistent
-# notifications. These calls would fail without this fixture since the persistent_notification
-# integration is never loaded during a test.
-@pytest.fixture(name="skip_notifications", autouse=True)
-def skip_notifications_fixture():
-    """Skip notification calls."""
-    with patch("homeassistant.components.persistent_notification.async_create"), patch(
-        "homeassistant.components.persistent_notification.async_dismiss"
-    ):
-        yield
-
-@pytest.fixture
-def entity_registry_enabled_by_default() -> Generator[AsyncMock, None, None]:
-    """Test fixture that ensures all entities are enabled in the registry."""
-    with patch(
-        "homeassistant.helpers.entity.Entity.entity_registry_enabled_default",
-        return_value=True,
-    ) as mock_entity_registry_enabled_by_default:
-        yield mock_entity_registry_enabled_by_default
 
 def readJson(fileName):
+    """Read filte to json."""
     test_filename = os.path.join(os.path.dirname(__file__), fileName)
     with open(test_filename, mode="rb") as json_file:
         return json.load(json_file)
 
 
 class MockPyViCare:
-    def __init__(self) -> None:
+    """Mocked PyVicare class based on a json dump."""
+
+    def __init__(self, fixture) -> None:
+        """Init a single device from json dump."""
         self.devices = [
-            PyViCareDeviceConfig(
-                ViCareServiceMock("fixtures/Vitodens300W.json"), "Vitodens", "online"
-            )
+            PyViCareDeviceConfig(ViCareServiceMock(fixture), "Vitodens", "online")
         ]
 
     def initWithCredentials(
         self, username: str, password: str, client_id: str, token_file: str
     ):
+        """Stub oauth login."""
         None
 
 
-def MockCircuitsData(circuits):
-    return {
-        "properties": {"enabled": {"value": circuits}},
-        "feature": "heating.circuits",
-    }
-
-
 class ViCareServiceMock:
+    """PyVicareService mock using a json dump."""
+
     def __init__(self, filename, rawInput=None):
+        """Initialize the mock from a json dump."""
         if rawInput is None:
             testData = readJson(filename)
             self.testData = testData
@@ -90,10 +59,12 @@ class ViCareServiceMock:
         self.setPropertyData = []
 
     def getProperty(self, property_name):
+        """Read a property from a json dump."""
         entities = self.testData["data"]
         return readFeature(entities, property_name)
 
     def setProperty(self, property_name, action, data):
+        """Set a property to its internal data structure."""
         self.setPropertyData.append(
             {
                 "url": buildSetPropertyUrl(self.accessor, property_name, action),
@@ -114,21 +85,8 @@ def mock_config_entry() -> MockConfigEntry:
 
 
 @pytest.fixture
-def mock_vicare() -> Generator[MagicMock, None, None]:
-    """Return a mocked ViCare client."""
-    with patch(
-        "homeassistant.components.vicare.vicare_login", return_value=MockPyViCare()
-    ), patch(
-        "PyViCare.PyViCareCachedService",
-        return_value=ViCareServiceMock("fixtures/Vitodens300W.json"),
-    ) as vicare_mock:
-        vicare = vicare_mock.return_value
-        yield vicare
-
-
-@pytest_asyncio.fixture
 async def init_integration(
-    hass: HomeAssistant, mock_config_entry: MockConfigEntry, mock_vicare: MagicMock
+    hass: HomeAssistant, mock_config_entry: MockConfigEntry
 ) -> MockConfigEntry:
     """Set up the ViCare integration for testing."""
     mock_config_entry.add_to_hass(hass)
@@ -137,3 +95,17 @@ async def init_integration(
     await hass.async_block_till_done()
 
     return mock_config_entry
+
+
+@pytest.fixture
+def mock_vicare_gas_boiler() -> Generator[MagicMock, None, None]:
+    """Return a mocked ViCare API representing a gas boiler device."""
+    with patch(
+        "homeassistant.components.vicare.vicare_login",
+        return_value=MockPyViCare("fixtures/Vitodens300W.json"),
+    ), patch(
+        "PyViCare.PyViCareCachedService",
+        return_value=ViCareServiceMock("fixtures/Vitodens300W.json"),
+    ) as vicare_mock:
+        vicare = vicare_mock.return_value
+        yield vicare
